# openapi.yaml
openapi: 3.0.3
info:
  title: Cup Stack Server API
  version: 0.1.0
  description: |
    Doosan M0609 컵 스태킹 로봇 제어 서버의 REST API.

    서비스별 포트:
    - **robot**: 8001
    - **handineye**: 8002
    - **handtoeye**: 8003

    > 이 파일은 참조용 문서입니다. 실제 Swagger UI는 FastAPI가 코드에서 자동 생성합니다.
    > 각 서비스의 `/docs` 엔드포인트에서 확인할 수 있습니다.
servers:
  - url: http://localhost:8001
    description: Robot service
  - url: http://localhost:8002
    description: Hand-in-eye calibration service
  - url: http://localhost:8003
    description: Hand-to-eye calibration service
...
paths:
...
  /api/robot/status:
    get:
      tags: [robot]
      summary: 로봇 상태 조회
      description: 관절 상태, 활성 태스크, bringup 상태, EE 위치, 전체 태스크 목록을 반환한다.
      responses:
        "200":
          description: 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RobotStatusResponse"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/robot/bringup:
    post:
      tags: [robot]
      summary: 로봇 bringup 시작
      description: |
        `mode=sim`이면 시뮬레이션 bringup 스크립트를 실행하고,
        `mode=real`이면 실로봇 bringup 스크립트를 `ip` 인자와 함께 실행한다.
        bringup 완료 20초 후 `move_cartesian` 서비스 노드가 자동 시작된다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BringupRequest"
            examples:
              sim:
                summary: 시뮬레이션
                value: { mode: sim }
              real:
                summary: 실로봇
                value: { mode: real, ip: "192.168.1.100" }
      responses:
        "200":
          description: bringup 프로세스 시작됨
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStartedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/robot/task/start:
    post:
      tags: [robot]
      summary: 태스크 시작
      description: |
        지정한 ROS 2 launch 태스크를 시작한다.
        이미 실행 중인 태스크가 있으면 409를 반환한다.
        `move_cartesian`은 서비스 노드이므로 액션 태스크와 동시에 실행 가능하다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskStartRequest"
            examples:
              gripper:
                summary: gripper 서비스 노드 시작
                value: { task: gripper, args: {} }
              cup_detection:
                summary: cup_detection 시작
                value: { task: cup_detection, args: {} }
              skill_api:
                summary: skill_api 서버 시작
                value: { task: skill_api, args: {} }
      responses:
        "200":
          description: 태스크 시작됨
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStartedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/robot/task/stop:
    post:
      tags: [robot]
      summary: 태스크 중지
      description: 실행 중인 태스크에 SIGINT를 보내고 최대 10초 후 SIGKILL로 강제 종료한다. 이미 정지된 태스크에 대해서는 즉시 반환한다(멱등).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskStopRequest"
            example:
              name: gripper
      responses:
        "200":
          description: 태스크 중지됨
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStoppedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/robot/task/log:
    get:
      tags: [robot]
      summary: 태스크 로그 조회
      description: 지정한 태스크의 최근 stdout 라인을 반환한다. MoveIt 노이즈 라인은 필터링된다.
      parameters:
        - name: name
          in: query
          required: true
          description: 태스크 이름
          schema:
            type: string
            example: gripper
        - name: tail
          in: query
          required: false
          description: 반환할 최대 줄 수
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 500
      responses:
        "200":
          description: 로그 반환
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskLogResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/robot/position:
    get:
      tags: [robot]
      summary: EE 위치 조회
      description: |
        마지막으로 명령한 엔드이펙터 위치(x, y, z)를 반환한다.
        `/api/robot/move`로 이동 명령을 내리기 전까지는 404가 반환된다.
      responses:
        "200":
          description: 마지막 명령 위치
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EEPosition"
              example:
                x: 0.35
                y: 0.0
                z: 0.40
        "404":
          description: 아직 이동 명령이 없어 위치 미확인
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/robot/config/workspace:
    get:
      tags: [robot]
      summary: 작업공간 한계 조회
      description: 현재 설정된 작업공간 좌표 한계 및 그리드 간격을 반환한다.
      responses:
        "200":
          description: 작업공간 한계
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceLimitsResponse"
              example:
                x_min: -0.5
                x_max: 0.5
                y_min: -0.5
                y_max: 0.5
                z_min: 0.25
                z_max: 0.55
                grid_spacing: 0.05
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/robot/config/pyramid:
    get:
      tags: [robot]
      summary: 피라미드 설정 조회
      description: |
        3-2-1 피라미드의 중심 좌표(cp), 행 방향 yaw(degree), pick Z 와
        6개 슬롯의 절대 place 좌표 캐시를 반환한다. cp 가 아직 설정되지
        않았으면 config 의 HOME XY 로 lazy 초기화한다.
      responses:
        "200":
          description: 현재 피라미드 설정
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PyramidConfigResponse"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    post:
      tags: [robot]
      summary: 피라미드 설정 갱신
      description: |
        center / degree / pick_z 중 보낸 필드만 갱신한다. degree 는
        `% 360` wrap, center XY 는 워크스페이스 범위 검증. 갱신 후 6개
        슬롯의 z 가 workspace z_max 를 초과하면 422.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PyramidConfigUpdate"
            examples:
              degree_only:
                summary: yaw 만 변경
                value: { degree: 45.0 }
              center_only:
                summary: center 만 변경
                value: { center: { x: 0.48, y: -0.05 } }
      responses:
        "200":
          description: 갱신된 피라미드 설정
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PyramidConfigResponse"
        "422":
          $ref: "#/components/responses/BadRequest"
  /api/robot/skill/pyramid:
    post:
      tags: [robot]
      summary: 피라미드 단위 스킬 (1 cup pick-and-place)
      description: |
        지정한 pick 좌표(x, y)에서 컵 하나를 잡아 피라미드의 지정된
        slot 위치에 놓는다. cp / degree / pick_z 는 `/config/pyramid` 에
        저장된 값을 자동 주입.

        호출자는 일반적으로 6번 (1l → 1m → 1r → 2l → 2r → 3m) 호출하여
        3-2-1 피라미드를 완성한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PyramidSkillRequest"
            examples:
              tier1_left:
                summary: "1l — bottom 왼쪽"
                value: { x: 0.40, y: 0.10, slot: "1l" }
              tier1_mid:
                summary: "1m — bottom 가운데"
                value: { x: 0.40, y: 0.10, slot: "1m" }
              tier1_right:
                summary: "1r — bottom 오른쪽"
                value: { x: 0.40, y: 0.10, slot: "1r" }
              tier2_left:
                summary: "2l — middle 왼쪽"
                value: { x: 0.40, y: 0.10, slot: "2l" }
              tier2_right:
                summary: "2r — middle 오른쪽"
                value: { x: 0.40, y: 0.10, slot: "2r" }
              tier3_top:
                summary: "3m — top (단일)"
                value: { x: 0.40, y: 0.10, slot: "3m" }
      responses:
        "200":
          description: 스킬 실행 결과
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PyramidSkillResponse"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/robot/skill/pick:
    post:
      tags: [robot]
      summary: 단일 컵 pick 스킬
      description: |
        지정한 컵 윗면 중앙(`x`, `y`)에서 컵 하나를 집는다. Z 는
        `z` (raw gripper Z) / `cup_top_z` (서버가 grip offset 가산) /
        `nested_count` (ROS 2 가 nested 스택 높이로 산출) 중 하나로 지정.
        ROS 2 `skill_api_node` 로 프록시되며 첫 호출 시 lazy 시작.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PickSkillRequest"
            examples:
              by_nested_count:
                summary: nested 스택 높이로 Z 산출
                value: { x: 0.45, y: -0.12, nested_count: 1 }
              by_cup_top_z:
                summary: 컵 윗면 Z 명시
                value: { x: 0.45, y: -0.12, cup_top_z: 0.05 }
              by_raw_z:
                summary: 그리퍼 raw Z 명시
                value: { x: 0.45, y: -0.12, z: 0.15 }
      responses:
        "200":
          description: 스킬 실행 결과
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PickSkillResponse"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/robot/skill/scan:
    post:
      tags: [robot]
      summary: 스캔 스킬 (양쪽 두 방향 후 초기 위치 복귀)
      description: |
        ROS 2 `skill_api_node` 의 `ScanSkill` 을 프록시한다.
        `scan.launch.py` 가 PTP 로 `초기 위치 → pos1 → pos2 → 초기 위치` 순으로
        이동하며 각 웨이포인트 도달 후 `dwell_sec`(기본 5초) 만큼 대기한다.
        인자가 없는 단일 스킬이므로 본문은 비어 있다.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
            example: {}
      responses:
        "200":
          description: 스킬 실행 결과
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanSkillResponse"
        "409":
          $ref: "#/components/responses/Conflict"
        "502":
          description: skill_api_node 가 HTTP 에러 반환
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/robot/cups:
    get:
      tags: [robot]
      summary: 컵 감지 결과 조회
      description: |
        ROS 2 `/cup_poses` 토픽의 최신 프레임 — YOLO 가 인식한 컵 목록과
        각 컵의 픽셀 좌표 · bbox · base_link 3D 위치(가능 시).
        `cup_detection` task 가 실행 중이어야 데이터가 들어온다.
      responses:
        "200":
          description: 마지막 감지 프레임
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CupDetectionFrame"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/robot/pixel-to-world:
    get:
      tags: [robot]
      summary: 픽셀 → base_link 좌표 변환
      description: |
        EE 장착 카메라(hand-in-eye)의 픽셀 좌표를 base_link 의 3D 좌표로
        변환한다. depth 이미지 · 카메라 intrinsics · hand-eye 캘리브레이션이
        모두 준비되어 있어야 한다.
      parameters:
        - name: px
          in: query
          required: true
          schema: { type: integer }
        - name: py
          in: query
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 변환된 좌표
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PixelToWorldResponse"
        "503":
          description: TF / intrinsics / depth / 캘리브레이션 미준비
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
  /api/robot/gripper:
    post:
      tags: [robot]
      summary: 그리퍼 제어
      description: |
        OnRobot RG2/RG6 그리퍼를 열거나 닫는다.
        `move_cartesian` 서비스 노드가 실행 중이어야 한다.
        하드웨어 미연결 시 `success=false`를 반환한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GripperRequest"
            examples:
              open:
                summary: 열기
                value: { command: open }
              close:
                summary: 닫기
                value: { command: close }
      responses:
        "200":
          description: 그리퍼 동작 결과
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GripperResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/robot/move:
    post:
      tags: [robot]
      summary: EE 이동
      description: |
        로봇 엔드이펙터를 지정한 좌표로 이동한다.
        `mode=absolute`이면 `base_link` 기준 절대 좌표, `mode=relative`이면 현재 위치에서의 델타.
        목표 좌표가 작업공간 한계를 벗어나면 자동으로 클램핑된다.
        `move_cartesian` 서비스 노드가 실행 중이어야 한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveRequest"
            examples:
              absolute:
                summary: 절대 좌표
                value: { x: 0.35, y: 0.0, z: 0.40, mode: absolute }
              relative:
                summary: 상대 이동
                value: { x: 0.05, y: 0.0, z: 0.0, mode: relative }
      responses:
        "200":
          description: 이동 결과
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoveResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  # ── Hand-in-Eye ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TaskStatus:
      type: string
      enum: [idle, running, stopping, failed]
      description: 태스크 프로세스 상태

    # ── Robot ──
    JointState:
      type: object
      properties:
        name:
          type: array
          items:
            type: string
          example: [joint_1, joint_2, joint_3, joint_4, joint_5, joint_6]
        position:
          type: array
          items:
            type: number
            format: double
          description: 라디안
        velocity:
          type: array
          items:
            type: number
            format: double
          description: 라디안/s
        effort:
          type: array
          items:
            type: number
            format: double
          description: N·m
...
    EEPosition:
      type: object
      required: [x, y, z]
      properties:
        x:
          type: number
          format: double
          description: base_link X (m)
        y:
          type: number
          format: double
          description: base_link Y (m)
        z:
          type: number
          format: double
          description: base_link Z (m)
    RobotStatusResponse:
      type: object
      required: [joints, task, bringup, tasks]
      properties:
        joints:
          $ref: "#/components/schemas/JointState"
        task:
          $ref: "#/components/schemas/ActiveTask"
        bringup:
          $ref: "#/components/schemas/ActiveTask"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/TaskSummary"
        ee_position:
          $ref: "#/components/schemas/EEPosition"
          nullable: true
    WorkspaceLimitsResponse:
      type: object
      required: [x_min, x_max, y_min, y_max, z_min, z_max, grid_spacing]
      properties:
        x_min:
          type: number
          format: double
          example: -0.5
        x_max:
          type: number
          format: double
          example: 0.5
        y_min:
          type: number
          format: double
          example: -0.5
        y_max:
          type: number
          format: double
          example: 0.5
        z_min:
          type: number
          format: double
          example: 0.25
        z_max:
          type: number
          format: double
          example: 0.55
        grid_spacing:
          type: number
          format: double
          example: 0.05
    BringupRequest:
      type: object
      required: [mode]
      properties:
        mode:
          type: string
          enum: [sim, real]
          default: sim
          description: "`real`이면 `ip` 필드 필요"
        ip:
          type: string
          default: "192.168.1.100"
          description: "`mode=real`일 때만 사용. 로봇 컨트롤러 IP"
          example: "192.168.1.100"
    TaskStartRequest:
      type: object
      required: [task]
      properties:
        task:
          type: string
          enum:
            - cup_detection
            - gripper
            - skill_api
            - bringup_sim
            - bringup_real
          description: |
            실행할 태스크 이름.

            | 값 | 설명 |
            |---|---|
            | `cup_detection` | YOLO 컵 인식 노드 (`/cup_poses` publish) |
            | `gripper` | OnRobot 그리퍼 ROS 서비스 노드 |
            | `skill_api` | pick / pyramid 단위 스킬 HTTP 서버 |
          example: gripper
        args:
          type: object
          additionalProperties:
            type: string
          description: |
            launch 파일에 전달할 추가 인자. `key:=value` 형식으로 변환된다.
          example: {}
    TaskStopRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: 중지할 태스크 이름
          example: gripper
    TaskStartedResponse:
      type: object
      required: [name, status]
      properties:
        name:
          type: string
          example: gripper
        status:
          $ref: "#/components/schemas/TaskStatus"
        pid:
          type: integer
          nullable: true
          example: 12346
    TaskStoppedResponse:
      type: object
      required: [name, status]
      properties:
        name:
          type: string
          example: gripper
        status:
          type: string
          example: stopped
    TaskLogResponse:
      type: object
      required: [name, log]
      properties:
        name:
          type: string
          example: gripper
        log:
          type: array
          items:
            type: string
          example:
            - "[INFO] [launch]: process started"
            - "[INFO] [gripper]: opening"
            - "[INFO] [gripper]: opened"
    GripperRequest:
      type: object
      required: [command]
      properties:
        command:
          type: string
          enum: [open, close]
          description: 그리퍼 동작 명령
    GripperResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
          example: "Gripper opened"
    MoveRequest:
      type: object
      required: [x, y, z]
      properties:
        x:
          type: number
          format: double
          description: base_link X (m)
        y:
          type: number
          format: double
          description: base_link Y (m)
        z:
          type: number
          format: double
          description: base_link Z (m)
        mode:
          type: string
          enum: [absolute, relative]
          default: absolute
    MoveResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        position:
          $ref: "#/components/schemas/EEPosition"
          nullable: true

    # ── Calibration ──
...
    PickSkillRequest:
      type: object
      required: [x, y]
      description: |
        좌표는 컵 윗면 중앙(base_link, m). Z 는 `z` (raw gripper Z) /
        `cup_top_z` (서버가 grip offset 가산) / `nested_count` (ROS 2 가
        nested 스택 높이로 산출) 중 정확히 하나로 지정.
      properties:
        x:
          type: number
          format: double
          description: 컵 윗면 중앙 X (base_link, m)
        y:
          type: number
          format: double
          description: 컵 윗면 중앙 Y (base_link, m)
        cup_top_z:
          type: number
          format: double
          nullable: true
          description: 컵 윗면 Z (m). 서버에서 grip offset 가산
        z:
          type: number
          format: double
          nullable: true
          description: 그리퍼 raw Z (m). cup_top_z 대신 직접 지정 시
        nested_count:
          type: integer
          minimum: 1
          nullable: true
          description: 소스 스택 nested 컵 개수. ROS 2 가 Z 산출
        ori:
          type: object
          nullable: true
          description: "그리퍼 방향 quaternion {x,y,z,w}. 미지정 시 down"
    PickSkillResponse:
      type: object
      required: [success, skill]
      properties:
        success: { type: boolean }
        skill: { type: string, example: pick }
        detail: { type: string }

    # ── Cup Detection ──
...
    CupDetectionFrame:
      type: object
      required: [stamp, frame_id, count, cups]
      properties:
        stamp:
          type: number
          format: double
          description: UNIX 타임스탬프 (초)
        frame_id:
          type: string
          example: base_link
        count:
          type: integer
          description: 감지된 컵 수
        cups:
          type: array
          items:
            $ref: "#/components/schemas/CupInfo"
    PixelToWorldResponse:
      type: object
      required: [x, y, z, depth_mm, pixel_x, pixel_y]
      properties:
        x: { type: number, format: double, description: base_link X (m) }
        y: { type: number, format: double, description: base_link Y (m) }
        z: { type: number, format: double, description: base_link Z (m) }
        depth_mm:
          type: integer
          description: 픽셀 깊이 (mm)
        pixel_x: { type: integer }
        pixel_y: { type: integer }

    # ── Pyramid Skill ──
...
    PyramidConfigResponse:
      type: object
      required: [center, degree, pick_z, slots]
      properties:
        center:
          $ref: "#/components/schemas/PyramidConfigCenter"
        degree:
          type: number
          format: double
          description: 행 방향 yaw (deg, [0, 360))
        pick_z:
          type: number
          format: double
          description: 그리퍼 pick Z (base_link, m)
        slots:
          type: object
          description: "키: 1l/1m/1r/2l/2r/3m, 값: 절대 place 좌표"
          additionalProperties:
            $ref: "#/components/schemas/PyramidSlotXYZ"
    PyramidConfigUpdate:
      type: object
      description: 보낸 필드만 갱신
      properties:
        center:
          $ref: "#/components/schemas/PyramidConfigCenter"
        degree:
          type: number
          format: double
        pick_z:
          type: number
          format: double
    PyramidSkillRequest:
      type: object
      required: [x, y, slot]
      properties:
        x:
          type: number
          format: double
          description: pick 컵 윗면 중앙 X (base_link, m)
        y:
          type: number
          format: double
          description: pick 컵 윗면 중앙 Y (base_link, m)
        slot:
          type: string
          enum: [ "1l", "1m", "1r", "2l", "2r", "3m" ]
          description: 놓을 슬롯 키
    PyramidSkillResponse:
      type: object
      required: [success, skill]
      properties:
        success: { type: boolean }
        skill: { type: string, example: pyramid }
        detail: { type: string }

    # ── Scan Skill ──
    ScanSkillResponse:
      type: object
      required: [success, skill]
      properties:
        success: { type: boolean }
        skill: { type: string, example: scan }
        detail: { type: string }

    # ── Error ──
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string
          description: 에러 설명
          example: "Invalid task. Choose from: ['bringup_real', 'bringup_sim', 'cup_detection', 'gripper', 'skill_api']"
  responses:
    BadRequest:
      description: 잘못된 요청 파라미터
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
          example:
            detail: "No task named 'gripper'"
    Conflict:
      description: 이미 실행 중인 태스크가 있어 새 태스크를 시작할 수 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
          example:
            detail: "Task 'bringup_real' is already running. Stop it first."
    ServiceUnavailable:
      description: 도메인이 초기화되지 않음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
          example:
            detail: "Robot domain not initialized"